refactor: remove deprecated data_processing.py and 01_filter_annotations.py
